commit tree b6f3973c0cc1536f11532e6568455a7ef5c32192

+++ app/controllers/toys_controller.rb
@@ -7,25 +7,33 @@ class ToysController < ApplicationController
   end
 
   def create
-    toy = Toys.create(toy_params)
+    toy = Toy.create(toy_params)
     render json: toy, status: :created
   end
 
   def update
     toy = Toy.find_by(id: params[:id])
-    toy.update(toy_params)
+    if toy
+      toy.update(toy_params)
+      render json: toy
+    else
+      render json: { error: "Toy not found"}, status: :not_found
+    end
   end
 
   def destroy
     toy = Toy.find_by(id: params[:id])
-    toy.destroy
-    head :no_content
+    if toy
+      toy.destroy
+      head :no_content
+    else
+      render json: { error: "Toy not found"}, status: :not_found
   end
 
   private
   
   def toy_params
-    params.permit(:name, :image, :likes)
+    params.permit(:name, :image, :likes, :id)
   end
 
 end
